fix: use int64_t for LongLongArray element (Linux LP64 build)
CI (ubuntu gcc) 报错：LongLongArray::AddElement 需要 int64_t*，
Linux 上 int64_t 是 long 而 long long 是不同类型，long long[3]
无法隐式转换为 int64_t*。改为 int64_t v[3] 跨平台均正确。

+++ Examples/Filter/TestExtractComponent.cpp
@@ -459,7 +459,7 @@ bool VerifyOutputTypePreserved() {
     vecLL->SetName("vecLL");
     vecLL->SetDimension(3);
     for (int i = 0; i < 4; ++i) {
-        long long v[3] = {10LL + i, 20LL + i, 30LL + i};
+        int64_t v[3] = {10LL + i, 20LL + i, 30LL + i};
         vecLL->AddElement(v);
     }
     meshLL->GetAttributeSet()->AddVector(IG_POINT, vecLL);
